+++ src/sections/orders/OrderGroup.js
@@ -1,73 +1,56 @@
-import React, { useState, useEffect } from "react";
-import {useDispatch} from "react-redux";
+import React, { useState } from "react";
 import { OrderView } from "./OrderView";
 import { TabsCustom } from "../../components/tabs";
 import { Tabs } from "antd"; /* Tharindu try to remove this part */
 
-import { isFetching } from "../../actions/common";
-import { getAllOrders } from "../../api/order";
-
 const { TabPane } = Tabs;
 
 export const OrderGroup = (props) => {
-  const { clickOK, refresh } = props;
-  const [allOrders, setAllOrders] = useState([]);
+  const { clickOK, typedOrders } = props;
   const [currentTab, setCurrentTab] = useState("1");
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    dispatch(isFetching(true));
-
-    getAllOrders().then((res) => {
-      if (res.data.status === "success") {
-        dispatch(isFetching(false));
-        setAllOrders(res.data.data);
-      }
-    });
-  }, [refresh]);
 
   const getCategoriesOrders = (tab) => {
     return (
-      allOrders &&
-      allOrders.filter((item) => {
+      typedOrders &&
+      typedOrders.filter((item) => {
         return item.status === tab;
       })
-      );
+    );
   };
 
   const tabCallback = (tab) => {
     // current tab can be get from currentTab if necessary
-    setCurrentTab(tab)
-  }
+    setCurrentTab(tab);
+  };
 
   return (
     <TabsCustom tabCallback={tabCallback}>
       <TabPane tab="Placed Orders" key="1">
-        <OrderView 
+        <OrderView
           clickOK={clickOK}
-          tab={"placed"} 
-          getTabOrders={getCategoriesOrders}
+          tab={"placed"}
+          getCategoriesOrders={getCategoriesOrders}
         />
       </TabPane>
       <TabPane tab="Prepared Orders" key="prepared">
-        <OrderView 
-          clickOK={clickOK} 
-          tab={"prepared"} 
-          getTabOrders={getCategoriesOrders}
+        <OrderView
+          clickOK={clickOK}
+          tab={"prepared"}
+          getCategoriesOrders={getCategoriesOrders}
         />
       </TabPane>
       <TabPane tab="Draft Orders" key="draft">
-        <OrderView 
-          clickOK={clickOK} 
-          tab={"draft"} 
-          getTabOrders={getCategoriesOrders}
+        <OrderView
+          clickOK={clickOK}
+          tab={"draft"}
+          getCategoriesOrders={getCategoriesOrders}
         />
       </TabPane>
       <TabPane tab="Settled Orders" key="settled">
-        <OrderView 
-          clickOK={clickOK} 
-          tab={"settled"} 
-          getTabOrders={getCategoriesOrders}
+        <OrderView
+          clickOK={clickOK}
+          tab={"settled"}
+          getCategoriesOrders={getCategoriesOrders}
         />
       </TabPane>
     </TabsCustom>
